% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalPhosPro.R
\name{CausalPhosPro}
\alias{CausalPhosPro}
\title{Return causal inference based on small sample size}
\usage{
CausalPhosPro(omicdata1, omicdata2, SNP, Priori, ratio, Covariates)
}
\arguments{
\item{omicdata1}{Exposure dataframe}

\item{omicdata2}{Outcome dataframe}

\item{Priori}{priori evidence}

\item{ratio}{SNP minor allele frequency}

\item{Covariates}{Common covariates include age, gender, and smoking status}
}
\value{
causal inference
}
\description{
The function returns a data frame with the IVW estimate, p-value, and confidence intervals
}
