% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis_snp_detector.R
\name{cis_snp_detector}
\alias{cis_snp_detector}
\title{Identifying cis-regulatory SNPs}
\usage{
cis_snp_detector(genelist, SNP, cisDist, GencodeAnnotation)
}
\value{
cis-regulatory SNPs
}
\description{
This function identifies cis-regulatory SNPs within a certain distance from genes of interest.
}
