% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phoslink.CI.R
\name{Phoslink.CI}
\alias{Phoslink.CI}
\title{Return causal inference based on small sample size}
\usage{
Phoslink.CI(omicdata1, omicdata2, SNP, Priori, ratio = 0.1, Covariates)
}
\arguments{
\item{omicdata1}{Exposure dataframe}

\item{omicdata2}{Outcome dataframe}

\item{Priori}{priori evidence}

\item{ratio}{SNP minor allele frequency}

\item{Covariates}{Common covariates include age, gender, and smoking status}
}
\value{
causal inference
}
\description{
The function returns a data frame with the IVW estimate, p-value, and confidence intervals
}
