% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cis_snp_detector.R
\name{cis_snp_detector}
\alias{cis_snp_detector}
\title{Identifying cis-regulatory SNPs}
\usage{
cis_snp_detector(gene, SNP, cisDist = 1e+06, GencodeAnnotation)
}
\arguments{
\item{gene}{A character vector containing the names of the genes of interest.}

\item{SNP}{A data frame containing SNP information, typically including chromosome, position.}

\item{cisDist}{An integer representing the distance threshold within which to search for Cis-regulatory SNPs, with a default value of 1e6 (1,000,000).}

\item{GencodeAnnotation}{A data frame containing gene annotation information, including chromosome, gene name, and location.}
}
\value{
cis-regulatory SNPs
}
\description{
This function identifies cis-regulatory SNPs within a certain distance from genes of interest.
}
